publisher/micropub/requirements.txt
boto3
mf2py
requests
multidict
dataclasses
sansio-multipart
unicode-slugify
pytz
python-dateutil